pyproject.toml
[tool.poetry]
name = "desert_rose"
version = "0.1.0"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
google-api-python-client = "^2.26.1"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
oauth2client = "^4.1.3"
requests = "^2.26.0"
google-cloud-bigquery = "^2.28.1"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^6.2.5"
pytest-timeout = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s --durations=0"
testpaths = [
    "test",
]
